requirements.txt
# Core ML & extraction
PyPDF2>=3.0.0
python-docx>=1.1.0
scikit-learn>=1.4.0

# New file type support
openpyxl>=3.1.0
python-pptx>=0.6.21
Pillow>=10.0.0

# Optional: OCR for images (requires Tesseract installed separately)
# pytesseract>=0.3.10

# Testing
pytest>=8.0.0
